- Added screen centering/reset with "r" key - confirmed connection to C++ client - Added captive civilian counter UI to Red Bases

+++ SOA/Assets/Custom Scripts/DataManager.cs
@@ -152,7 +152,7 @@ namespace soa
                 Debug.Log("Adding actor to actor dictionary " + actor.unique_id);
                 soaActorDictionary[actor.unique_id] = actor;
                 actorDistanceDictionary[actor.unique_id] = new SortedDictionary<int,bool>();
-                addBelief(new Belief_Actor(actor.unique_id, actor.affiliation, actor.type,
+                addBelief(new Belief_Actor(actor.unique_id, (int)actor.affiliation, actor.type,
                     actor.displayPosition.x, actor.displayPosition.y, actor.displayPosition.z), actor.unique_id);
             }
             else
